pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel", "setuptools-scm>=8"]   # scm optional but nice
build-backend = "setuptools.build_meta"

[project]
name = "idh-prediction"
version = "0.1.0"                        # or omit if using setuptools-scm for tags
description = "IDH Prediction Project"
authors = [{name = "User"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.3,<2.4",
  "pyarrow>=16,<22",
  "google-cloud-aiplatform>=1.120,<2",
  "google-cloud-bigquery>=3.38,<4",
  "plotly>=5,<6", # 6.x is very new; pin major if you need stability
  "python-dotenv>=1.0,<2",
  "gradio>=5.0,<6", # or remove: pandas will pull a compatible numpy
  "numpy>=1.26",
  "pytest>=8.4.2",
  "gcsfs>=2025.9.0",
  "db-dtypes>=1.4.3",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "pre-commit",
  "pip-tools",                           # if you choose pip-tools
]
serve = [
  "uvicorn",
]

[tool.setuptools]
package-dir = {"" = "src"}               # canonical src-layout

[tool.setuptools.packages.find]
where = ["src"]
include = ["idh*"]

# Type checking & lint
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E","F","I","UP","B","SIM"]
